Add Script to Generate Portable Python Function to Load DuckDB Export
This is something I can use in the generated blog going forward to enable running a pre-loaded duckdb instance in a Marimo Snippet using the new marimo-snippets lib that just got released! That integration will come next.

This script is exciting because I can finally make DuckDB "portable" within entirely WASM Marimo notebooks by just running an initial build step to export the duckdb database and run this script over it to generate something that should translate over to WASM Marimo just fine.

There's still a bit of a bummer in the fact that for my specific use case it looks like I'll still need to hack out the HNSW index since the vss DuckDB extension can't be loaded from Pyodide's DuckDB version just yet :/. So I won't be able to use `<=>` to do the most interesting vizualizations over clustering etc, I'd have to use numpy or something. It's not the end of the world.