Marimo Integration in Generated Blog Post!
Turns out the script I previously made was actually not able to work with the Marimo Snippets integration since the way the snippets library works is it just packs up all the code into a URI encoding and sends it to `https://marimo.app?code=<encoded code>`. Instead I was able to keep tinkering and figured out that apparently the marimo env allows you to write to files.... that's absolutely wild.... but it lets me load the notes database persisted to GitHub and write it to a file in the Marimo env, which effectively works around the fact that WASM duckdb (in Pyodide) doesn't support ATTACHing to remote database files. This is a sweet integration! At this point the page is cluttered as all heck, but it's a really sweet demo.